# services/workflow-control/docs/api/runner-openapi.yaml
openapi: 3.1.0
info:
  title: OpenSlack Workflow Runner Control API
  version: 1.1.0-gs9g
  description: >-
    Private default-off runner lifecycle service. GS9-F2b provides the
    loopback-only authority-binding companion for exact checkpoint, effect,
    budget, and resume delivery. GS9-G permits the same exact v2 route only for
    immutable, explicitly allowlisted new-record canary receipts after durable
    Workflow Control accept. Legacy and rollback TS routes remain on v1.
servers:
  - url: http://127.0.0.1:8081
security: []
...
paths:
  /v1/runner/jobs:
    post:
      operationId: submitWorkflowRunnerJob
      description: >-
        Admit one exact canonical, hash-only job specification. The idempotency
        key and request fingerprint must bind the canonical request exactly.
      x-openslack-canonical-json: required
      x-openslack-go-authority: [job, receipt]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/JobIdempotencyKey'
        - $ref: '#/components/parameters/RequestFingerprint'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobSpec' }
      responses:
        '201': { $ref: '#/components/responses/JobAccepted' }
        '200': { $ref: '#/components/responses/JobDuplicate' }
        '202': { $ref: '#/components/responses/JobReconciliation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '408': { $ref: '#/components/responses/Timeout' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v2/runner/jobs:
    post:
      operationId: submitWorkflowRunnerV2QualificationJob
      description: >-
        Default-off GS9-F qualification or GS9-G explicit new-record canary
        admission. The exact body permanently binds protocol v2, required
        capabilities, authority route, run revision, and resume generation.
      x-openslack-canonical-json: required
      x-openslack-availability: qualification-or-new-record-canary
      x-openslack-no-protocol-downgrade: true
      x-openslack-replay-response: exact-original
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/JobIdempotencyKeyV2'
        - $ref: '#/components/parameters/RequestFingerprint'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/V2JobSpec' }
      responses:
        '201':
          description: V2 qualification or new-record canary job accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/V2JobReceipt' }
        '200':
          description: Exact accepted or reconciliation-required v2 receipt replayed byte-for-byte.
          headers:
            Idempotency-Replayed:
              schema: { const: 'true' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/V2JobReceipt' }
        '202':
          description: V2 admission requires reconciliation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/V2JobReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '408': { $ref: '#/components/responses/Timeout' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v2/runner/authority-bindings:stage:
    post:
      operationId: stageWorkflowRunnerAuthorityBinding
      description: >-
        Qualification or new-record canary companion sequence 1. Durably
        reserve the exact future runner event before its source authority is
        committed.
      x-openslack-canonical-json-lf: required
      x-openslack-availability: qualification-or-new-record-canary
      x-openslack-runtime-delivery-enable-required: true
      x-openslack-loopback-only: true
      x-openslack-replay-response: exact-original
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/BindingIdempotencyKey'
        - $ref: '#/components/parameters/RequestFingerprint'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthorityBindingStage' }
      responses:
        '201': { $ref: '#/components/responses/AuthorityBindingAccepted' }
        '200': { $ref: '#/components/responses/AuthorityBindingDuplicate' }
        '202': { $ref: '#/components/responses/AuthorityBindingReconciliation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '408': { $ref: '#/components/responses/Timeout' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v2/runner/runtime-admissions:seal:
    post:
      operationId: sealWorkflowRunnerV2RuntimeAdmission
      description: >-
        Qualification or new-record canary durable admission between the
        sealed worker's recovery-projection disposition point-read and its
        lease_accept event. It distinguishes an initial generation-zero run
        from the first resume without changing frozen runner-v2 or F2a bytes.
      x-openslack-canonical-json-lf: required
      x-openslack-availability: qualification-or-new-record-canary
      x-openslack-runtime-delivery-enable-required: true
      x-openslack-loopback-only: true
      x-openslack-replay-response: exact-original
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/RuntimeAdmissionIdempotencyKey'
        - $ref: '#/components/parameters/RequestFingerprint'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/V2RuntimeAdmission' }
      responses:
        '201':
          description: Exact runtime disposition sealed before lease acceptance.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/V2RuntimeAdmissionReceipt' }
        '200':
          description: Exact original runtime-admission receipt replayed byte-for-byte.
          headers:
            Idempotency-Replayed:
              schema: { const: 'true' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/V2RuntimeAdmissionReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '408': { $ref: '#/components/responses/Timeout' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v2/runner/authority-bindings/receipts/{idempotencyKey}:
    get:
      operationId: getWorkflowRunnerAuthorityBindingReceipt
      description: >-
        Point-read the exact durable companion receipt after a lost POST
        response. A missing receipt never authorizes replay under another key.
      x-openslack-availability: qualification-or-new-record-canary
      x-openslack-runtime-delivery-enable-required: true
      x-openslack-loopback-only: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BindingReceiptKey'
        - $ref: '#/components/parameters/WorkspaceHeader'
      responses:
        '200':
          description: Exact original durable authority-binding receipt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthorityBindingReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/runner/jobs/{jobId}:
    get:
      operationId: getWorkflowRunnerJob
      x-openslack-go-authority: [job, attempt, lease, fence, receipt]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobID'
        - $ref: '#/components/parameters/WorkspaceHeader'
      responses:
        '200':
          description: Current runner lifecycle projection for the bound job.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/runner/jobs/{jobId}/cancellations:
    post:
      operationId: cancelWorkflowRunnerJob
      description: >-
        Admit a cancellation bound to the exact current attempt, lease, fence,
        correlation, request time, and expiry. Exact replay returns the original
        accepted response with the same status and bytes.
      x-openslack-canonical-json: required
      x-openslack-go-authority: [cancel, fence, receipt]
      x-openslack-replay-response: exact-original
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobID'
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/CancelIdempotencyKey'
        - $ref: '#/components/parameters/RequestFingerprint'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelRequest' }
      responses:
        '202':
          description: Original durable cancellation admission receipt, including on exact replay.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CancelReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '408': { $ref: '#/components/responses/Timeout' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-random-token
      description: Raw bearer value is verified against the configured SHA-256 identity.
  parameters:
    WorkspaceHeader:
      name: X-OpenSlack-Workspace-ID
      in: header
      required: true
      description: Must equal the one workspace sealed into this runner-server process.
      schema: { $ref: '#/components/schemas/Identifier' }
    JobID:
      name: jobId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Identifier' }
    JobIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-job\.v1\.[0-9a-f]{64}$'
        maxLength: 160
    JobIdempotencyKeyV2:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-job\.v2\.[0-9a-f]{64}$'
        maxLength: 160
    CancelIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-cancel\.v1\.[0-9a-f]{64}$'
        maxLength: 160
...
    BindingReceiptKey:
      name: idempotencyKey
      in: path
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-authority-binding\.v1\.[0-9a-f]{64}$'
        maxLength: 160
    BindingIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-authority-binding\.v1\.[0-9a-f]{64}$'
        maxLength: 160
    RuntimeAdmissionIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        pattern: '^openslack\.workflow-runner-v2-runtime-admission\.v1\.[0-9a-f]{64}$'
        maxLength: 160
    RequestFingerprint:
      name: X-OpenSlack-Request-Fingerprint
      in: header
      required: true
      schema:
        type: string
        pattern: '^sha256:[0-9a-f]{64}$'
        maxLength: 71
  responses:
    JobAccepted:
      description: Job and original receipt committed atomically.
      content: { application/json: { schema: { $ref: '#/components/schemas/JobReceipt' } } }
    JobDuplicate:
      description: Original durable job receipt returned for exact replay.
      content: { application/json: { schema: { $ref: '#/components/schemas/JobReceipt' } } }
    JobReconciliation:
      description: Submission commit outcome requires durable reconciliation.
      content: { application/json: { schema: { $ref: '#/components/schemas/JobReceipt' } } }
    AuthorityBindingAccepted:
      description: Exact authority-binding phase committed durably.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthorityBindingReceipt' }
    AuthorityBindingDuplicate:
      description: Exact original authority-binding receipt replayed byte-for-byte.
      headers:
        Idempotency-Replayed:
          schema: { const: 'true' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthorityBindingReceipt' }
    AuthorityBindingReconciliation:
      description: Exact durable receipt requires reconciliation; no source/event replay is authorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthorityBindingReceipt' }
    Unauthorized:
      description: Bearer or single-workspace identity is invalid.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Timeout:
      description: Bounded request deadline exceeded.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency, identity, lease, or fencing precondition conflicted.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooLarge:
      description: >-
        Exact request exceeds its route-specific frozen limit: 64 KiB for job,
        runtime-admission, and cancellation requests; 1 MiB for authority-binding frames.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    UnsupportedMediaType:
      description: Content-Type is not exactly application/json.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Closed shape, canonical bytes, hash, or header binding is invalid.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: The bound runner job, authority binding, or idempotency receipt does not exist.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Internal:
      description: Sanitized internal integrity failure.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unavailable:
      description: PostgreSQL or commit-outcome recovery is unavailable.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Identifier:
      type: string
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:@-]{0,255}$'
      maxLength: 256
    Hash:
      type: string
      pattern: '^[0-9a-f]{64}$'
      minLength: 64
      maxLength: 64
    Timestamp:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$'
      minLength: 24
      maxLength: 24
    SafePositiveInteger:
      type: integer
      format: int64
      minimum: 1
      maximum: 9007199254740991
    Semver:
      type: string
      pattern: '^(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?$'
      minLength: 5
      maxLength: 64
    JobSpec:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required:
        [
          schema,
          workspaceId,
          jobId,
          workflowRunId,
          correlationId,
          executionDescriptorRef,
          executionDescriptorHash,
          workflowId,
          workflowVersion,
          workflowSourceHash,
          manifestHash,
          inputHash,
          wholeTimeoutMs,
          submittedAt,
        ]
      properties:
        schema: { const: openslack.workflow_runner_job_spec.v1 }
        workspaceId: { $ref: '#/components/schemas/Identifier' }
        jobId: { $ref: '#/components/schemas/Identifier' }
        workflowRunId: { $ref: '#/components/schemas/Identifier' }
        correlationId: { $ref: '#/components/schemas/Identifier' }
        executionDescriptorRef: { $ref: '#/components/schemas/Identifier' }
        executionDescriptorHash: { $ref: '#/components/schemas/Hash' }
        workflowId: { $ref: '#/components/schemas/Identifier' }
        workflowVersion: { $ref: '#/components/schemas/Identifier' }
        workflowSourceHash: { $ref: '#/components/schemas/Hash' }
        manifestHash: { $ref: '#/components/schemas/Hash' }
        inputHash: { $ref: '#/components/schemas/Hash' }
        wholeTimeoutMs:
          allOf:
            - $ref: '#/components/schemas/SafePositiveInteger'
            - minimum: 1000
              maximum: 86400000
        submittedAt: { $ref: '#/components/schemas/Timestamp' }
...
    V2JobSpec:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required:
        [
          schema,
          workspaceId,
          jobId,
          workflowRunId,
          correlationId,
          executionDescriptorRef,
          executionDescriptorHash,
          workflowId,
          workflowVersion,
          workflowSourceHash,
          manifestHash,
          inputHash,
          wholeTimeoutMs,
          submittedAt,
          requiredProtocolVersion,
          requiredCapabilities,
          authorityRoute,
          runRevision,
          resumeGeneration,
        ]
      properties:
        schema: { const: openslack.workflow_runner_job_spec.v2 }
        workspaceId: { $ref: '#/components/schemas/Identifier' }
        jobId: { $ref: '#/components/schemas/Identifier' }
        workflowRunId: { $ref: '#/components/schemas/Identifier' }
        correlationId: { $ref: '#/components/schemas/Identifier' }
        executionDescriptorRef: { $ref: '#/components/schemas/Identifier' }
        executionDescriptorHash: { $ref: '#/components/schemas/Hash' }
        workflowId: { $ref: '#/components/schemas/Identifier' }
        workflowVersion: { $ref: '#/components/schemas/Semver' }
        workflowSourceHash: { $ref: '#/components/schemas/Hash' }
        manifestHash: { $ref: '#/components/schemas/Hash' }
        inputHash: { $ref: '#/components/schemas/Hash' }
        wholeTimeoutMs:
          allOf:
            - $ref: '#/components/schemas/SafePositiveInteger'
            - minimum: 1000
              maximum: 86400000
        submittedAt: { $ref: '#/components/schemas/Timestamp' }
        requiredProtocolVersion: { const: openslack.workflow_runner.v2 }
        requiredCapabilities:
          type: array
          prefixItems:
            - { const: cancel_ack }
            - { const: effect_receipts }
            - { const: lease_heartbeat }
          minItems: 3
          maxItems: 3
        authorityRoute: { $ref: '#/components/schemas/AuthorityRouteV2' }
        runRevision: { $ref: '#/components/schemas/SafePositiveInteger' }
        resumeGeneration:
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740991
    V2JobReceipt:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      oneOf:
        - title: AcceptedV2JobReceipt
          properties:
            status: { const: accepted }
            state: { const: queued }
            revision: { const: 1 }
            reconciliationId: { type: 'null' }
        - title: ReconciliationRequiredV2JobReceipt
          properties:
            status: { const: reconciliation_required }
            state: { const: reconciliation_required }
            reconciliationId:
              type: string
              minLength: 1
              maxLength: 512
      required:
        [
          schema,
          status,
          workspaceId,
          jobId,
          workflowRunId,
          state,
          revision,
          jobSpecHash,
          idempotencyKey,
          requestFingerprint,
          committedAt,
          reconciliationId,
        ]
      properties:
        schema: { const: openslack.workflow_runner_job_receipt.v2 }
        status: { enum: [accepted, reconciliation_required] }
        workspaceId: { $ref: '#/components/schemas/Identifier' }
        jobId: { $ref: '#/components/schemas/Identifier' }
        workflowRunId: { $ref: '#/components/schemas/Identifier' }
        state: { enum: [queued, reconciliation_required] }
        revision: { $ref: '#/components/schemas/SafePositiveInteger' }
        jobSpecHash: { $ref: '#/components/schemas/Hash' }
        idempotencyKey:
          type: string
          pattern: '^openslack\.workflow-runner-job\.v2\.[0-9a-f]{64}$'
        requestFingerprint:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        committedAt: { $ref: '#/components/schemas/Timestamp' }
        reconciliationId:
          type: [string, 'null']
          maxLength: 512
    V2RuntimeAdmission:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required:
        [
          schema,
          workspaceId,
          jobId,
          workflowRunId,
          attemptId,
          leaseId,
          fencingToken,
          jobSpecHash,
          disposition,
        ]
      properties:
        schema: { const: openslack.workflow_runner_v2_runtime_admission.v1 }
        workspaceId: { $ref: '#/components/schemas/Identifier' }
        jobId: { $ref: '#/components/schemas/Identifier' }
        workflowRunId: { $ref: '#/components/schemas/Identifier' }
        attemptId: { $ref: '#/components/schemas/Identifier' }
        leaseId: { $ref: '#/components/schemas/Identifier' }
        fencingToken: { $ref: '#/components/schemas/SafePositiveInteger' }
        jobSpecHash: { $ref: '#/components/schemas/Hash' }
        disposition: { enum: [initial, resume] }
    V2RuntimeAdmissionReceipt:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required:
        [
          schema,
          status,
          workspaceId,
          jobId,
          workflowRunId,
          attemptId,
          leaseId,
          fencingToken,
          jobSpecHash,
          disposition,
          idempotencyKey,
          requestFingerprint,
          committedAt,
        ]
      properties:
        schema: { const: openslack.workflow_runner_v2_runtime_admission_receipt.v1 }
        status: { const: accepted }
        workspaceId: { $ref: '#/components/schemas/Identifier' }
        jobId: { $ref: '#/components/schemas/Identifier' }
        workflowRunId: { $ref: '#/components/schemas/Identifier' }
        attemptId: { $ref: '#/components/schemas/Identifier' }
        leaseId: { $ref: '#/components/schemas/Identifier' }
        fencingToken: { $ref: '#/components/schemas/SafePositiveInteger' }
        jobSpecHash: { $ref: '#/components/schemas/Hash' }
        disposition: { enum: [initial, resume] }
        idempotencyKey:
          type: string
          pattern: '^openslack\.workflow-runner-v2-runtime-admission\.v1\.[0-9a-f]{64}$'
        requestFingerprint:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        committedAt: { $ref: '#/components/schemas/Timestamp' }
...
    CancelRequest:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required:
        [
          schema,
          correlationId,
          expectedAttemptId,
          expectedLeaseId,
          expectedFence,
          reason,
          requestedAt,
          expiresAt,
        ]
      properties:
        schema: { const: openslack.workflow_runner_cancel_request.v1 }
        correlationId: { $ref: '#/components/schemas/Identifier' }
        expectedAttemptId: { $ref: '#/components/schemas/Identifier' }
        expectedLeaseId: { $ref: '#/components/schemas/Identifier' }
        expectedFence: { $ref: '#/components/schemas/SafePositiveInteger' }
        reason: { enum: [operator, lease_expired, shutdown, superseded, timeout] }
        requestedAt: { $ref: '#/components/schemas/Timestamp' }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
...
    Error:
      type: object
      additionalProperties: false
      unevaluatedProperties: false
      required: [schema, code, message]
      properties:
        schema: { const: openslack.workflow_runner_control_error.v1 }
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{0,127}$'
          maxLength: 128
        message: { type: string, minLength: 1, maxLength: 512 }